Controller/Machine.py: pass the peer name when moving device to netns
In particular for veth device pair, when first device from the pair is
moved to netns, the peer name is generated from the device name with
additional index suffix that is incorrect. For example for the devicepair of
{lveth0, peer_lveth0}, when lveth0 is moved to a namespace, the peer is
set to peer_lveth00.

To fix this, we can pass the peer name directly in the remap_device
call.

Signed-off-by: Jan Tluka <jtluka@redhat.com>